Refactor file structure and styles for Shelf Editor
