+++ src/editors/ShelfEditor/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './ShelfEditor.less';
+import './style.less';
 
 class ShelfEditor extends React.Component {
   constructor(props) {
@@ -71,38 +71,27 @@ class ShelfEditor extends React.Component {
   render() {
     let ActionBar = this.props.actionBar;
     return (
-      <div className="v-shelf-ed">
-        <form className="v-shelf-ed__inner">
-          <div className="v-shelf-ed__text-field">
+      <div className="ShelfEditor">
+        <form className="ShelfEditor__inner" onSubmit={this.handleSubmit.bind(this)}>
+          <div className="ShelfEditor__text-field">
             <label htmlFor="shelf-title">Título da Prateleira</label>
             <input id="shelf-title" className="form-control" name="title" type="text"
                    value={this.state.title} onChange={this.changeValue.bind(this)}
                    placeholder="Ex: Destaques, Promoção"/>
           </div>
-
-          <div className="v-shelf-ed__text-field">
+          <div className="ShelfEditor__text-field">
             <label htmlFor="shelf-category">Categoria</label>
             <input id="shelf-category" className="form-control" name="category" type="text"
                    value={this.state.category} onChange={this.changeValue.bind(this)}
                    placeholder="Ex: camiseta, cadeira"/>
           </div>
-
-          <div className="v-shelf-ed__text-field">
-            <label htmlFor="shelf-collection">Coleção</label>
-            <input id="shelf-colletion" className="form-control" name="collection" type="text"
-                   value={this.state.collection} onChange={this.changeValue.bind(this)}
-                   placeholder="Ex: 001, 002"/>
-          </div>
-
-          <div className="v-shelf-ed__quant-field">
+          <div className="ShelfEditor__quant-field">
             <label htmlFor="shelf-quantity">Quantidade de Produtos na Prateleira</label>
-            <div className="v-shelf-ed__quant-selector">
-              <button className="v-shelf-ed__quant-button--left" onClick={this.decrementQuantity.bind(this)}> - </button>
-
-                <input id="shelf-quantity" className="v-shelf-ed__quant-selector__input" name="quantity" type="text"
+            <div className="ShelfEditor__quant-selector">
+              <button className="ShelfEditor__quant-button--left" onTouchTap={this.decrementQuantity.bind(this)}> - </button>
+                <input id="shelf-quantity" className="ShelfEditor__quant-selector__input" name="quantity" type="text"
                        value={this.state.quantity} onChange={this.changeQuantity.bind(this)}/>
-
-              <button className="v-shelf-ed__quant-button--right" onClick={this.incrementQuantity.bind(this)}> + </button>
+              <button className="ShelfEditor__quant-button--right" onTouchTap={this.incrementQuantity.bind(this)}> + </button>
             </div>
           </div>
         </form>
